requirements.txt
# Core dependencies
matplotlib>=3.5.0
numpy>=1.20.0
scipy>=1.7.0
Pillow>=9.0.0
pyserial>=3.5
mpmath>=1.2.0
